+++ internal/parser/toml_validation.go
@@ -3,9 +3,11 @@ package parser
 import (
 	"fmt"
 	"regexp"
+
+	"github.com/snapp-incubator/barat/internal/config"
 )
 
-func TomlValidation(tomlFiles map[string]map[string]interface{}, excludeRegex []string) (errs []error) {
+func TomlValidation(tomlFiles map[string]map[string]interface{}) (errs []error) {
 	checkedKeys := map[string]struct{}{}
 
 	for lang, tomls := range tomlFiles { // lang = en, ru, etc. tomls = map[string]interface{}
@@ -14,7 +16,7 @@ func TomlValidation(tomlFiles map[string]map[string]interface{}, excludeRegex []
 			_, ok := checkedKeys[key]
 			if !ok {
 				for lan, _ := range tomlFiles {
-					if !isExcluded(key, excludeRegex) {
+					if !isExcluded(key, config.C.Exclude.RegexKey) {
 						if _, ok := tomlFiles[lan][key]; !ok {
 							errs = append(errs,
 								fmt.Errorf("key \"%s\" not found in tag \"%s\"", key, lan))
@@ -25,24 +27,28 @@ func TomlValidation(tomlFiles map[string]map[string]interface{}, excludeRegex []
 			}
 
 			// check if value is valid for all tags or not
-			if d, ok := value.(map[string]interface{})["description"]; ok {
-				if d == "" {
+			if config.C.Options.Enable.DescriptionCheck {
+				if d, ok := value.(map[string]interface{})["description"]; ok {
+					if d == "" {
+						errs = append(errs,
+							fmt.Errorf("description key is empty: key \"%s\", tag \"%s\"", key, lang))
+					}
+				} else {
 					errs = append(errs,
-						fmt.Errorf("description key is empty: key \"%s\", tag \"%s\"", key, lang))
+						fmt.Errorf("description key not found: key \"%s\", tag \"%s\"", key, lang))
 				}
-			} else {
-				errs = append(errs,
-					fmt.Errorf("description key not found: key \"%s\", tag \"%s\"", key, lang))
 			}
 
-			if o, ok := value.(map[string]interface{})["other"]; ok {
-				if o == "" {
+			if config.C.Options.Enable.OtherKeyCheck {
+				if o, ok := value.(map[string]interface{})["other"]; ok {
+					if o == "" {
+						errs = append(errs,
+							fmt.Errorf("other key is empty: key \"%s\", tag \"%s\" ", key, lang))
+					}
+				} else {
 					errs = append(errs,
-						fmt.Errorf("other key is empty: key \"%s\", tag \"%s\" ", key, lang))
+						fmt.Errorf("other key not found: key \"%s\", tag \"%s\"", key, lang))
 				}
-			} else {
-				errs = append(errs,
-					fmt.Errorf("other key not found: key \"%s\", tag \"%s\"", key, lang))
 			}
 		}
 	}