+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 
 * Fix duplicate reports bug for missing keys in the toml files. (2023-01-26, @n25a, !5)
-
+* Support config file as input for checker command. (2021-01-27, @n25a, !6)
 
 ## v0.0.1